feat: enhance music bot text channel creation and permissions
This commit enhances the creation of the music bot text channel by changing its name to "bytebot-🎵" and placing it in the same category as the context channel. It also updates the permissions set for the default role in this channel, disallowing sending messages, managing messages, embedding links, attaching files, mentioning everyone, using external emojis, and managing threads.

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>