Dev (#9)
* refactor: ensure channel_id is not none during music setup

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* refactor: update db stuff

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* refactor: ensure queue is being updated when a song is added

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* fix: update queue and volume fields correctly in embed

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* refactor: streamline music control buttons and fix defer response

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* style: change search button style to secondary

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* feat: enhance music bot text channel creation and permissions

This commit enhances the creation of the music bot text channel by changing its name to "bytebot-🎵" and placing it in the same category as the context channel. It also updates the permissions set for the default role in this channel, disallowing sending messages, managing messages, embedding links, attaching files, mentioning everyone, using external emojis, and managing threads.

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* chore: ignore commitmsgbase.txt in git

This commit updates the .gitignore file to ignore commitmsgbase.txt. This prevents the file from being tracked by git, keeping the repository clean from temporary or unneeded files.

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

* fix: correct SQL query in music bot code

This commit corrects an SQL query in the music bot code. The query was incorrectly trying to fetch the 'music_message_id' from the 'GuildSettings' table, but it has been corrected to fetch the 'music_channel_id'. This ensures that the correct data is retrieved from the database.

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>

---------

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>