fix: update message edit calls in music.py
This commit addresses an issue in the 'on_wavelink_track_start' and 'on_wavelink_track_end' methods in the 'Music' cog. Previously, the 'message.edit' method was not being called with the 'view' parameter, which could have caused the message's components to not update correctly. This commit ensures that the 'view' parameter is included in the 'message.edit' calls, which should allow the message's components to update as expected.

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>