fix: create new discord.ui.View instance in music.py
This commit addresses an issue in the 'on_wavelink_track_start' and 'on_wavelink_track_end' methods in the 'Music' cog. Previously, the 'message.edit' method was being called with a list of components, which caused an AttributeError. This commit ensures that a new discord.ui.View instance is created and the components are added to it before calling the 'message.edit' method. This should prevent the AttributeError from occurring.

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>