fix: add ActionRow instances to discord.ui.View in music.py
This commit addresses an issue in the 'on_wavelink_track_start' and 'on_wavelink_track_end' methods in the 'Music' cog. Previously, the 'discord.ui.View.add_item()' method was being called with instances of 'discord.ui.Button', which caused a TypeError. This commit ensures that 'discord.ui.ActionRow' instances are added to the 'discord.ui.View' instance. This should prevent the TypeError from occurring.

Signed-off-by: Lu <42706009+luinbytes@users.noreply.github.com>